+++ apps/dashboard/app/components/cards/overview/index.tsx
@@ -1,0 +1,5 @@
+import OverviewCardGrid from "./grid";
+import type { OverviewCardProps } from "./card";
+
+export default OverviewCardGrid;
+export type { OverviewCardProps };
